Split version-related logic into another package
